some linux updates

+++ inc/btle/characteristic.h
@@ -54,7 +54,7 @@ namespace btle {
 
         uint8_t properties() const;
         const btle::uuid& uuid() const;
-        uint8_t attribute_handle() const;
+        uint16_t attribute_handle() const;
         long int instance_id() const;
         btle::service* parent();
